WIP #38 - Renamed related fixtures (breaking tests first)
