+++ fixtures/innerclasses/SC2.java
